Remove deprecated and unused `WithPublicDashboardEndpoints`
